Split out Request handling into an Extension. Dup the Application (like Sinatra) when responding
